Discover node ID from KubeVirt
harvester-csi-driver previously assumed the Harvester VM's name was the
hostname, but this is not necessarily always the case. When the cluster
operator creates the VM and specifically makes the hostname different
from the VM name, harvester-csi-driver would not be able to select the
correct VMI based on just the hostname.

To fix this, discover the VMI name by comparing the MAC addresses seen
from within the guest to the MAC addresses described in the VMI.

Signed-off-by: Connor Kuehl <connor.kuehl@suse.com>